+++ methods/matchingnet.py
@@ -178,7 +178,7 @@ class MatchingNet(MetaTemplate):
 
         # Get the query labels
         # y_query = self.get_episode_labels(self.n_query, enable_grad=True)
-        y_query = y[:, self.n_support :].reshape(self.n_way * self.n_query)
+        y_query = y[:, self.n_support :].reshape(self.n_way * self.n_query).long()
 
         # Compute the scores (logprobs)
         outputs = self.set_forward(x)
