+++ methods/protonet.py
@@ -170,7 +170,7 @@ class ProtoNet(MetaTemplate):
 
         # Get the query labels
         # y_query = self.get_episode_labels(self.n_query, enable_grad=True)
-        y_query = y[:, self.n_support :].reshape(self.n_way * self.n_query)
+        y_query = y[:, self.n_support :].reshape(self.n_way * self.n_query).long()
 
         # Compute the scores
         outputs = self.set_forward(x)
